libs/wd_tagger/requirements.txt
numpy
onnxruntime
pandas
pillow